+++ tests/dev_and_build/as_js_classic_library/js_classic_library_dev.test_manual.mjs
@@ -3,6 +3,10 @@ import { startDevServer } from "@jsenv/core"
 startDevServer({
   logLevel: "info",
   rootDirectoryUrl: new URL("./client/", import.meta.url),
-  clientAutoreload: false,
+  clientAutoreload: true,
+  clientFiles: {
+    "./*/**": true,
+  },
   supervisor: false,
+  clientMainFileUrl: new URL("./client/main.html", import.meta.url),
 })